+++ changelog.md
@@ -5,6 +5,8 @@
 - Added `//schem delete`
 - Implemented configurable settings via `worldedit.properties`
   - Added `//reloadconfig` to reload configuration file
+- Added metadata to schematics
+  - Schematics now keep track of required mods
 
 ## v0.4.0+7.3_04